+++ CHANGELOG.md
@@ -8,6 +8,8 @@ that will be sent.
 - The functions `get_and_send_messages`and `send_msg can accept string or bytes.
 - Removed function `receive_msg`.
 - Connection errors and socket timeout are logged with different messages.
+- `Client`, `ClientSender`, `ServerSender` and `Server` have `send_wait` attribute to optionally wait some time
+before sending new messages.
 
 ### Tests
 - Unit test for send and receive modules.